requirements.txt
absl-py==2.0.0
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==0.5.2
annotated-types==0.6.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blessed==1.20.0
braceexpand==0.1.7
cachetools==5.3.1
carvekit-colab==4.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
clip==1.0
cmake==3.27.6
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.21.0
decorator==4.4.2
depth-diff-gaussian-rasterization-min==0.0.0
diffusers==0.30.0
dill==0.3.8
easydict==1.11
einops==0.7.0
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.103.2
fastjsonschema==2.18.1
filelock==3.12.4
Flask==2.2.5
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.1
fvcore==0.1.5.post20221221
google-auth==2.23.2
google-auth-oauthlib==1.0.0
gorilla-core==0.2.7.8
gpustat==1.1.1
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.24.6
hydra-core==1.3.2
icecream==2.1.3
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.7
imgaug==0.4.0
imhist==0.0.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
imWatermark==0.0.2
iniconfig==2.0.0
iopath==0.1.10
ipython==8.12.3
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kornia==0.7.0
latent-diffusion @ file:///data/pug/RGBD-Diffusion/third_party/latent-diffusion
lazy_loader==0.3
lightning-utilities==0.9.0
lit==17.0.2
lmdb==1.4.1
loguru==0.7.2
lpips==0.1.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nbformat==5.7.0
nerfacc==0.5.3
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1
numpy==1.24.4
nvdiffrast @ file:///data/pug/RGBD-Diffusion/third_party/nvdiffrast
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.0.2
open3d==0.17.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
peft==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
piq==0.8.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.17.0
pluggy==1.3.0
plyfile==1.0.3
portalocker==2.8.2
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyMCubes==0.1.4
pymeshlab==2022.2.post4
pynvml==11.5.0
pyparsing==3.1.1
pyquaternion==0.9.9
pyre-extensions==0.0.23
pytest==7.4.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.9.0
# Editable install with no version control (pytorch-ssim==0.1)
-e /home/pug/.conda/envs/perf/lib/python3.8/site-packages
# Editable install with no version control (pytorch3d==0.7.4)
-e /data/pug/pytorch3d-main
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
recon @ file:///data/pug/RGBD-Diffusion/third_party/recon
referencing==0.30.2
regex==2023.10.3
requests==2.32.3
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.6.0
roma==1.4.2
rpds-py==0.10.4
rsa==4.9
Rtree==1.1.0
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
shapely==2.0.1
simple-knn==0.0.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
taming-transformers-rom1504==0.0.6
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.8.10.dev0
tinycudann==1.7
tokenizers==0.14.1
tomli==2.0.1
torch==2.0.0
torch-efficient-distloss==0.1.3
torch-fidelity==0.3.0
torchmetrics==1.2.0
torchvision==0.15.2+cu117
tqdm==4.66.5
traitlets==5.11.2
transformers==4.34.0
trimesh==3.23.5
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
wcwidth==0.2.8
webdataset==0.2.57
Werkzeug==2.2.3
widgetsnbextension==4.0.9
xformers==0.0.17
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
